nfc: obscure association ids from user
I meant to do this a while ago... Oops
Basically, we keep a ring buffer with 8 entries, containing
sha256(association_id + ":" + game_id) so that association_id <=> tag
one-to-one relationship is retained without the security concerns. The
ring buffer ensures we don't grow unbounded, the game_id component of
the hash ensures privacy for association IDs so they can't be tracked
across games. The reason for the 8 entry capacity was to make sure
multiple users wouldn't get lost if a game was doing something funky.
I'd love some opinions though :)